Remove shared library
